User Authorization

+++ app/Http/Controllers/ListingController.php
@@ -57,6 +57,12 @@ class ListingController extends Controller
 
     //update listing
     public function update(Request $request, Listing $listing){
+
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -78,6 +84,12 @@ class ListingController extends Controller
 
     //delete listing
     public function destroy(Listing $listing) {
+
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message','Listing deleted successfully');
     }
